requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
Flask==2.2.3
frozenlist==1.3.3
google-search-results==2.4.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.134
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.0
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
soupsieve==2.4
SQLAlchemy==1.4.47
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wikipedia==1.4.0
yarl==1.8.2
